chore : add pessimistic write lock on event ticketprice to lock the row from updates chore : add reserve method in booking service

+++ src/main/java/com/tiko/tiko/Events/DTO/UpdateEventRequest.java
@@ -8,6 +8,6 @@ public record UpdateEventRequest(
         EventStatus status,
         String location,
         LocalDateTime date,
-        Long capacity,
+        int capacity,
         String imageUrl
 ) {}
